chatbot_service/requirements.txt
flask
flask-cors
pymongo
openai
google-generativeai
python-dotenv
numpy
pandas
tiktoken
chromadb
pymongo[encryption]
motor